pyproject.toml
[project]
name = "douyin-spider-web"
version = "1.0.0"
description = "抖音UP主视频查询Web应用"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "requests>=2.31.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "loguru>=0.7.0",
    "ujson>=5.9.0",
    "python-dotenv>=1.0.0",
    "pymysql>=1.1.0",
    "sqlalchemy>=2.0.0",
    "apscheduler>=3.10.0",
    "gmssl>=3.2.2",
    "cryptography>=46.0.5",
    "httpx>=0.28.1",
    "zhipuai>=2.1.5.20250825",
    "sniffio>=1.3.1",
    "dashscope>=1.25.13",
    "paramiko>=4.0.0",
    "oss2>=2.19.1",
    "websocket-client>=1.9.0",
    "pywhispercpp>=1.4.1",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-core>=0.3.0",
    "python-jose[cryptography]>=3.5.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.22",
    "bcrypt>=5.0.0",
    "slowapi>=0.1.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "httpx>=0.26.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["backend"]
